test(all): refactor tests & add first test for find-the-expert page
Closes #210
